ai-coding-course-claude/lessons/lesson-02-multi-file-editing/src/claude-2/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claude-2"
version = "0.1.0"
description = "Transcript analyzer application with Gemini integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai",
    "pydantic",
    "python-dotenv",
]

[project.scripts]
main = "src.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]